requirements.txt
# Image libraries
Pillow
opencv-python
# PyTorch
torch
torchvision
# Other
matplotlib
numpy
tqdm